feat: add round settlement hook and tests
